refactor(player): 사용되지 않는 강제 형변환 삭제

+++ src/main/java/domain/model/Player.java
@@ -56,7 +56,7 @@ public class Player implements WithDeck {
     }
 
     public int getFinalMoney() {
-        return (int)bet.getFinalMoney(playerStatus);
+        return bet.getFinalMoney(playerStatus);
     }
 
     // 추가 카드 부여
